+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- **CLI**: Remove `pico-args`
+
 ## [0.15.0] - 2025-06-17
 
 ### Changed