+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 
 - **CLI**: Remove `pico-args`
+- Bump MSRV to `1.75`.
+- Update `selectors` to `0.29`.
 
 ## [0.15.0] - 2025-06-17