+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 - Update `selectors` to `0.29`.
 - Update `html5ever` to `0.35`.
 
+### Performance
+
+- Use interned string to compare `style` element name.
+
 ## [0.15.0] - 2025-06-17
 
 ### Changed