+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 - **CLI**: Remove `pico-args`
 - Bump MSRV to `1.75`.
-- Update `selectors` to `0.29`.
+- Update `selectors` to `0.30`.
 - Update `html5ever` to `0.35`.
 
 ### Performance